casestudy/requirements.txt
python-dotenv
pinecone
pypdf
langchain
langchain-community
langchain-openai
langchain-pinecone
streamlit